extractor/requirements.txt
iniconfig==2.0.0
lxml==4.9.3
packaging==23.2
Pillow==9.5.0
pluggy==1.3.0
pytest==7.4.2
python-pptx==0.6.22
XlsxWriter==3.1.9
